[WIP] Experiment with reusing `EvaluationContext` across runs
